requirements.txt
pandas
mysql-connecter-python
requests